+++ lib/elasticsearch_ex/api/document.ex
@@ -92,7 +92,7 @@ defmodule ElasticsearchEx.Api.Document do
   end
 
   def index(source, index, document_id, opts)
-      when is_map(source) and is_index(index) and is_document_id(document_id) do
+      when is_map(source) and is_index(index) and is_identifier(document_id) do
     Client.put("/#{index}/_doc/#{document_id}", nil, source, opts)
   end
 
@@ -134,7 +134,7 @@ defmodule ElasticsearchEx.Api.Document do
   @doc since: "1.0.0"
   @spec create(source(), index(), document_id(), opts()) :: ElasticsearchEx.response()
   def create(source, index, document_id, opts \\ [])
-      when is_map(source) and is_index(index) and is_document_id(document_id) do
+      when is_map(source) and is_index(index) and is_identifier(document_id) do
     Client.put("/#{index}/_create/#{document_id}", nil, source, opts)
   end
 
@@ -173,7 +173,7 @@ defmodule ElasticsearchEx.Api.Document do
   @doc since: "1.0.0"
   @spec get(index(), document_id(), opts()) :: ElasticsearchEx.response()
   def get(index, document_id, opts \\ [])
-      when is_index(index) and is_document_id(document_id) do
+      when is_index(index) and is_identifier(document_id) do
     Client.get("/#{index}/_doc/#{document_id}", nil, nil, opts)
   end
 
@@ -225,7 +225,7 @@ defmodule ElasticsearchEx.Api.Document do
   def get_ids(document_ids, index, opts)
       when is_list(document_ids) and is_index(index) do
     Enum.each(document_ids, fn document_id ->
-      is_document_id(document_id) ||
+      is_identifier(document_id) ||
         raise ArgumentError, "invalid value, expected a binary, got: `#{inspect(document_id)}`"
     end)
 
@@ -375,7 +375,7 @@ defmodule ElasticsearchEx.Api.Document do
       Enum.all?(values, &is_map/1) ->
         get_docs(values, index, opts)
 
-      Enum.all?(values, &is_document_id/1) ->
+      Enum.all?(values, &is_identifier/1) ->
         get_ids(values, index, opts)
 
       true ->
@@ -400,7 +400,7 @@ defmodule ElasticsearchEx.Api.Document do
   @doc since: "1.0.0"
   @spec exists?(index(), document_id(), opts()) :: boolean()
   def exists?(index, document_id, opts \\ [])
-      when is_index(index) and is_document_id(document_id) do
+      when is_index(index) and is_identifier(document_id) do
     Client.head("/#{index}/_doc/#{document_id}", nil, opts) == :ok
   end
 
@@ -439,7 +439,7 @@ defmodule ElasticsearchEx.Api.Document do
   @doc since: "1.0.0"
   @spec delete(index(), document_id(), opts()) :: ElasticsearchEx.response()
   def delete(index, document_id, opts \\ [])
-      when is_index(index) and is_document_id(document_id) do
+      when is_index(index) and is_identifier(document_id) do
     Client.delete("/#{index}/_doc/#{document_id}", nil, nil, opts)
   end
 
@@ -483,7 +483,7 @@ defmodule ElasticsearchEx.Api.Document do
   @doc since: "1.0.0"
   @spec update(source(), index(), document_id(), opts()) :: ElasticsearchEx.response()
   def update(source, index, document_id, opts \\ [])
-      when is_map(source) and is_index(index) and is_document_id(document_id) do
+      when is_map(source) and is_index(index) and is_identifier(document_id) do
     Client.post("/#{index}/_update/#{document_id}", nil, source, opts)
   end
 end