+++ lib/elasticsearch_ex/api/document/source.ex
@@ -44,7 +44,7 @@ defmodule ElasticsearchEx.Api.Document.Source do
   @doc since: "1.0.0"
   @spec get(index(), document_id(), keyword()) :: ElasticsearchEx.response()
   def get(index, document_id, opts \\ [])
-      when is_index(index) and is_document_id(document_id) do
+      when is_index(index) and is_identifier(document_id) do
     Client.get("/#{index}/_source/#{document_id}", nil, nil, opts)
   end
 
@@ -64,7 +64,7 @@ defmodule ElasticsearchEx.Api.Document.Source do
   @doc since: "1.0.0"
   @spec exists?(index(), document_id(), keyword()) :: boolean()
   def exists?(index, document_id, opts \\ [])
-      when is_index(index) and is_document_id(document_id) do
+      when is_index(index) and is_identifier(document_id) do
     Client.head("/#{index}/_source/#{document_id}", nil, opts) == :ok
   end
 end