+++ test/elasticsearch_ex/guards_test.exs
@@ -3,77 +3,46 @@ defmodule ElasticsearchEx.GuardsTest do
 
   import ElasticsearchEx.Guards
 
-  describe "is_index/1" do
-    test "with nil" do
-      refute is_index(nil)
-    end
-
-    test "with an atom" do
-      assert is_index(:index)
-    end
-
-    test "with a string" do
-      assert is_index("index")
-    end
-
-    test "with an empty string" do
-      refute is_index("")
-    end
+  test "is_index_or_nil/1" do
+    assert is_index_or_nil(nil)
+    assert is_index_or_nil(:index)
+    assert is_index_or_nil("index")
+    refute is_index_or_nil("")
+    refute is_index_or_nil([])
+    refute is_index_or_nil(%{})
+    refute is_index_or_nil({})
+    refute is_index_or_nil(%Stream{})
   end
 
-  describe "is_document_id/1" do
-    test "with nil" do
-      refute is_document_id(nil)
-    end
-
-    test "with an atom" do
-      refute is_document_id(:doc_id)
-    end
-
-    test "with a string" do
-      assert is_document_id("doc_id")
-    end
-
-    test "with an empty string" do
-      refute is_document_id("")
-    end
+  test "is_index/1" do
+    refute is_index(nil)
+    assert is_index(:index)
+    assert is_index("index")
+    refute is_index("")
+    refute is_index([])
+    refute is_index(%{})
+    refute is_index({})
+    refute is_index(%Stream{})
   end
 
-  describe "is_enum/1" do
-    test "with list" do
-      assert is_enum([])
-    end
-
-    test "with a Stream" do
-      stream = Stream.map([], & &1)
-
-      assert is_enum(stream)
-    end
-
-    test "with a map" do
-      refute is_enum(%{})
-    end
-
-    test "with nil" do
-      refute is_enum(nil)
-    end
+  test "is_identifier/1" do
+    refute is_identifier(nil)
+    refute is_identifier(:doc_id)
+    assert is_identifier("doc_id")
+    refute is_identifier("")
+    refute is_identifier([])
+    refute is_identifier(%{})
+    refute is_identifier({})
+    refute is_identifier(%Stream{})
   end
 
-  describe "is_identifier/1" do
-    test "with nil" do
-      refute is_identifier(nil)
-    end
-
-    test "with an atom" do
-      assert is_identifier(:doc_id)
-    end
-
-    test "with a string" do
-      assert is_identifier("doc_id")
-    end
-
-    test "with an empty string" do
-      refute is_identifier("")
-    end
+  test "is_enum/1" do
+    assert is_enum([])
+    assert is_enum(%Stream{})
+    refute is_enum(nil)
+    refute is_enum(:atom)
+    refute is_enum("string")
+    refute is_enum(%{})
+    refute is_enum({})
   end
 end